refactor: move the new helpers to a separate file
